refactoring credentials as a shared package.
