+++ internal/sdk/credentials/credentials.go
@@ -1,24 +1,37 @@
-package hooks
+// Package credentials provides shared Cribl credential loading from environment
+// and ~/.cribl/credentials. Used by the SDK hooks and the import CLI.
+package credentials
 
 import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"gopkg.in/ini.v1"
 	"log"
 	"os"
 	"path/filepath"
+
+	"gopkg.in/ini.v1"
 )
 
+// CriblConfig holds Cribl auth and connection settings (env or file).
+type CriblConfig struct {
+	ClientID       string `json:"client_id" ini:"client_id"`
+	ClientSecret   string `json:"client_secret" ini:"client_secret"`
+	OrganizationID string `json:"organization_id" ini:"organization_id"`
+	Workspace      string `json:"workspace" ini:"workspace"`
+	CloudDomain    string `json:"cloud_domain" ini:"cloud_domain"`
+
+	OnpremServerURL string `json:"onprem_server_url" ini:"onprem_server_url"`
+	OnpremUsername  string `json:"onprem_username" ini:"onprem_username"`
+	OnpremPassword  string `json:"onprem_password" ini:"onprem_password"`
+}
+
 func checkLocalConfigDir() ([]byte, error) {
-	// Then try ~/.cribl/credentials file
 	homeDir, err := os.UserHomeDir()
 	if err != nil {
 		log.Printf("[ERROR] Failed to get home directory: %v", err)
 		return []byte{}, fmt.Errorf("failed to get home directory: %v", err)
 	}
-
-	// Check for credentials in ~/.cribl/credentials
 	configDir := filepath.Join(homeDir, ".cribl")
 	configPath := filepath.Join(configDir, "credentials")
 	var filePath string
@@ -32,9 +45,8 @@ func checkLocalConfigDir() ([]byte, error) {
 			if errors.Is(err, os.ErrNotExist) {
 				log.Printf("[DEBUG] No config file found %s", legacyPath)
 				return []byte{}, err
-			} else {
-				return []byte{}, err
 			}
+			return []byte{}, err
 		}
 		filePath = legacyPath
 	} else {
@@ -46,41 +58,32 @@ func checkLocalConfigDir() ([]byte, error) {
 	if err != nil {
 		return []byte{}, fmt.Errorf("failed to read credentials file: %v", err)
 	}
-
 	return file, nil
 }
 
 func checkConfigFileFormat(input []byte) (string, error) {
 	var data interface{}
-	err := json.Unmarshal(input, &data)
-	if err == nil {
+	if err := json.Unmarshal(input, &data); err == nil {
 		return "json", nil
 	}
-
-	_, err = ini.Load(input)
-	if err == nil {
+	if _, err := ini.Load(input); err == nil {
 		return "ini", nil
 	}
-
 	return "", errors.New("config file type not recognized")
 }
 
 func parseJSONConfig(file []byte) (*CriblConfig, error) {
 	log.Printf("[DEBUG] parsing JSON config")
-
 	var config CriblConfig
-	err := json.Unmarshal(file, &config)
-	if err != nil {
+	if err := json.Unmarshal(file, &config); err != nil {
 		log.Printf("[ERROR] Failed to parse config file: %v", err)
 		return nil, fmt.Errorf("failed to parse config file: %v", err)
 	}
-
 	return &config, nil
 }
 
 func parseIniConfig(file []byte) (*CriblConfig, error) {
 	config := CriblConfig{}
-
 	profileName := os.Getenv("CRIBL_PROFILE")
 	if profileName == "" {
 		profileName = "default"
@@ -92,28 +95,23 @@ func parseIniConfig(file []byte) (*CriblConfig, error) {
 		log.Printf("[ERROR] Failed to parse config file: %v", err)
 		return nil, fmt.Errorf("failed to parse config file: %v", err)
 	}
-
-	err = cfg.Section(profileName).MapTo(&config)
-	if err != nil {
+	if err := cfg.Section(profileName).MapTo(&config); err != nil {
 		log.Printf("[ERROR] Failed to parse config file profile: %v", err)
 		return nil, fmt.Errorf("failed to parse config file profile: %v", err)
 	}
-
 	log.Printf("[DEBUG] Selected profile values - clientID=%s, orgID=%s, workspace=%s, domain=%s",
 		config.ClientID, config.OrganizationID, config.Workspace, config.CloudDomain)
 	return &config, nil
 }
 
-// GetCredentials reads credentials from environment variables or credentials file
+// GetCredentials reads credentials from environment variables or ~/.cribl/credentials.
+// Env takes precedence; file is used when env is not set.
 func GetCredentials() (*CriblConfig, error) {
-	// First try environment variables
 	clientID := os.Getenv("CRIBL_CLIENT_ID")
 	clientSecret := os.Getenv("CRIBL_CLIENT_SECRET")
 	organizationID := os.Getenv("CRIBL_ORGANIZATION_ID")
 	workspace := os.Getenv("CRIBL_WORKSPACE_ID")
 	cloudDomain := os.Getenv("CRIBL_CLOUD_DOMAIN")
-
-	// Read on-prem environment variables
 	onpremServerURL := os.Getenv("CRIBL_ONPREM_SERVER_URL")
 	onpremUsername := os.Getenv("CRIBL_ONPREM_USERNAME")
 	onpremPassword := os.Getenv("CRIBL_ONPREM_PASSWORD")
@@ -121,7 +119,6 @@ func GetCredentials() (*CriblConfig, error) {
 	log.Printf("[DEBUG] Environment variables - clientID=%s, orgID=%s, workspace=%s, domain=%s, onpremURL=%s",
 		clientID, organizationID, workspace, cloudDomain, onpremServerURL)
 
-	// If we have direct credentials in environment, use them
 	if clientID != "" && clientSecret != "" {
 		log.Printf("[DEBUG] Using cloud credentials from environment variables")
 		return &CriblConfig{
@@ -132,8 +129,6 @@ func GetCredentials() (*CriblConfig, error) {
 			CloudDomain:    cloudDomain,
 		}, nil
 	}
-
-	// If we have on-prem credentials in environment, use them
 	if onpremServerURL != "" {
 		log.Printf("[DEBUG] Using on-prem credentials from environment variables")
 		return &CriblConfig{
@@ -154,9 +149,8 @@ func GetCredentials() (*CriblConfig, error) {
 				Workspace:      workspace,
 				CloudDomain:    cloudDomain,
 			}, nil
-		} else {
-			return nil, err
 		}
+		return nil, err
 	}
 
 	format, err := checkConfigFileFormat(file)
@@ -164,13 +158,11 @@ func GetCredentials() (*CriblConfig, error) {
 		log.Printf("[DEBUG] No configuration file found, continuing - error: %v", err)
 		return nil, err
 	}
-
 	switch format {
 	case "json":
 		return parseJSONConfig(file)
 	case "ini":
 		return parseIniConfig(file)
 	}
-
 	return nil, nil
 }
