+++ internal/sdk/internal/hooks/types.go
@@ -3,6 +3,8 @@ package hooks
 import (
 	"sync"
 	"time"
+
+	credpkg "github.com/criblio/terraform-provider-criblio/internal/sdk/credentials"
 )
 
 type TokenInfo struct {
@@ -33,18 +35,8 @@ type CriblTerraformHook struct {
 	workspaceID string
 }
 
-type CriblConfig struct {
-	ClientID       string `json:"client_id" ini:"client_id"`
-	ClientSecret   string `json:"client_secret" ini:"client_secret"`
-	OrganizationID string `json:"organization_id" ini:"organization_id"`
-	Workspace      string `json:"workspace" ini:"workspace"`
-	CloudDomain    string `json:"cloud_domain" ini:"cloud_domain"`
-
-	// On-prem configuration fields
-	OnpremServerURL string `json:"onprem_server_url" ini:"onprem_server_url"`
-	OnpremUsername  string `json:"onprem_username" ini:"onprem_username"`
-	OnpremPassword  string `json:"onprem_password" ini:"onprem_password"`
-}
+// CriblConfig is the shared credential config from internal/sdk/credentials.
+type CriblConfig = credpkg.CriblConfig
 
 type CriblConfigFile struct {
 	Profiles map[string]CriblConfig `json:"profiles"`
